fix(charts): keep VictoryChartContext out of VictoryChartPie
Move chart context usage into VictoryChartPieLabel (prod API) so polar
pie charts work without prop-drilling label config through the slice tree.
